+++ src/Network/Algorand/Node/Api.hs
@@ -1,0 +1,107 @@
+-- SPDX-FileCopyrightText: 2021 Serokell <https://serokell.io>
+--
+-- SPDX-License-Identifier: MPL-2.0
+
+-- | The REST API v2 of Algod, the Algorand node.
+--
+-- See <https://developer.algorand.org/docs/reference/rest-apis/algod/v2/>
+module Network.Algorand.Node.Api
+  ( Api (..)
+  , ApiAny (..)
+  , ApiV2 (..)
+
+  , BuildVersion (..)
+  , Version (..)
+  , Account (..)
+  ) where
+
+import GHC.Generics (Generic)
+
+import Data.Aeson.TH (deriveJSON)
+import Data.Int (Int64)
+import Data.Text (Text)
+import Data.Word (Word64)
+import Servant.API ((:>), Capture, Get, JSON)
+import Servant.API.Generic (ToServantApi, (:-))
+
+import Data.Algorand.Address (Address)
+import Data.Algorand.Amount (Microalgos)
+import Network.Algorand.Node.Api.Json (algorandSnakeOptions, algorandTrainOptions)
+
+
+-- | Node software build version information.
+data BuildVersion = BuildVersion
+  { bvBranch :: Text
+  , bvBuildNumber :: Int64
+  , bvChannel :: Text
+  , bvCommitHash :: Text
+  , bvMajor :: Int64
+  , bvMinor :: Int64
+  }
+  deriving (Generic, Show)
+$(deriveJSON algorandSnakeOptions 'BuildVersion)
+
+-- | algod version information.
+data Version = Version
+  { vBuild :: BuildVersion
+  , vGenesisHashB64 :: Text
+  , vGenesisId :: Text
+  , vVersions :: [Text]
+  }
+  deriving (Generic, Show)
+$(deriveJSON algorandSnakeOptions 'Version)
+
+
+data Account = Account
+  { aAddress :: Address
+  , aAmount :: Microalgos
+  , aAmountWithoutPendingRewards :: Microalgos
+  --, aAppsLocalState :: Maybe
+  --, aAppsTotalSchema :: Maybe
+  --, aAssets :: Maybe
+  , aAuthAddr :: Maybe Address
+  --, aCreatedApps :: Maybe
+  --, aCreatedAssets :: Maybe
+  --, aAccuntParticipation :: Maybe
+  , aPendingRewards :: Microalgos
+  , aRewardBase :: Maybe Microalgos
+  , aRewards :: Microalgos
+  , aRound :: Word64
+  --, aSigType :: Maybe
+  , aStatus :: Text
+  }
+  deriving (Generic, Show)
+$(deriveJSON algorandTrainOptions 'Account)
+
+
+-- | The part of the API that does not depend on the version.
+data ApiAny route = ApiAny
+  { _health :: route
+      :- "health"
+      :> Get '[JSON] ()
+  , _version :: route
+      :- "versions"
+      :> Get '[JSON] Version
+  }
+  deriving (Generic)
+
+
+-- | Algod API (v2 only).
+data ApiV2 route = ApiV2
+  { _account :: route
+      :- "accounts"
+      :> Capture "address" Address
+      :> Get '[JSON] Account
+  }
+  deriving (Generic)
+
+
+-- | Algod API.
+data Api route = Api
+  { _vAny :: route
+      :- ToServantApi ApiAny
+  , _v2 :: route
+      :- "v2"
+      :> ToServantApi ApiV2
+  }
+  deriving (Generic)